+++ Source/Utils/HangarConfig.cs
@@ -5,34 +5,21 @@
 //
 //  Copyright (c) 2015 Allis Tauri
 
-using System;
-using System.Linq;
-using System.Collections.Generic;
-using UnityEngine;
+using System.Diagnostics.CodeAnalysis;
 using AT_Utils;
+using UnityEngine;
 
 namespace AtHangar
 {
-    /// <summary>
-    /// Loads hangar configuration presets at game loading
-    /// </summary>
-    [KSPAddon(KSPAddon.Startup.Instantly, true)]    
-    public class HangarGlobalsLoader : MonoBehaviour
-    { public void Start() { Globals.Load(); } }
-
-    class Globals : PluginGlobals<Globals>
+    [SuppressMessage("ReSharper", "FieldCanBeMadeReadOnly.Global"),
+     SuppressMessage("ReSharper", "ConvertToConstant.Global"),
+     SuppressMessage("ReSharper", "MemberCanBePrivate.Global")]
+    internal class Globals : PluginGlobals<Globals>
     {
         //store vessel
         [Persistent] public float  MaxSqrRelVelocity     = 1f;    //m/s
         [Persistent] public float  MaxSqrRelAcceleration = 0.01f; //m/s2
         [Persistent] public bool   EnableVesselPacking   = true;
-        //restore vessel
-        [Persistent] public float  MaxSqrAngularVelocity = 0.01f; //5.73 deg/s
-        [Persistent] public float  MaxSqrSurfaceVelocity = 0.01f; //m/s
-        [Persistent] public float  MaxGeeForce           = 0.2f;  //g
-        [Persistent] public float  MaxStaticPressure     = 0.01f; //atm
-        //misc
-        [Persistent] public string KethaneMapCollider  = "MapOverlay collider";
         [Persistent] public bool   UseStockAppLauncher = false;
         [Persistent] public string DontCloneResources  = "ElectricCharge, LiquidFuel, Oxidizer, Ore, XenonGas, MonoPropellant";
         public string[] ResourcesBlacklist { get; private set; }